rag_backed/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
numpy
python-multipart
openai
python-dotenv